backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
pydantic==2.6.1
qdrant-client>=1.7.3
requests==2.31.0
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.2.0
ftfy>=6.1.1
regex>=2023.12.25
tqdm>=4.66.1
git+https://github.com/openai/CLIP.git 